staging: cxt1e1: fix pointer-integer size mismatch warning
Fix the pointer-integer size mismatch warning below:
	drivers/staging/cxt1e1/functions.c: In function ‘VMETRO_TRACE’:
	drivers/staging/cxt1e1/functions.c:268:21: warning: cast from pointer
			to integer of different size [-Wpointer-to-int-cast]
	     u_int32_t   y = (u_int32_t) x;
			     ^

+++ drivers/staging/cxt1e1/functions.c
@@ -265,7 +265,7 @@ extern ci_t *CI;                /* dummy pointer to board ZERO's data */
 void
 VMETRO_TRACE (void *x)
 {
-    u_int32_t   y = (u_int32_t) x;
+    u_int32_t   y = (u_int32_t)(uintptr_t) x;
 
     pci_write_32 ((u_int32_t *) &CI->cpldbase->leds, y);
 }
